+++ changelog.md
@@ -55,6 +55,7 @@
 * example 可从 CMakeLists.private.txt 文件中加载 components 目录
 * 完善了 JSDoc 文档
 * EventEmitter::emitNativeEvent() 增加参数 bool fromISR=false 支持从 ISR 中触发
+* repl.enableUnknownPkgEvent() / repl.disableUnknownPkgEvent() — 监听 pkg.unknown 事件
 
 # v1.0.10